requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
apscheduler
supabase
loguru
python-multipart
